requirements.txt
git+https://github.com/snarfed/arroba.git#egg=arroba
git+https://github.com/snarfed/carbox.git#egg=carbox
git+https://github.com/snarfed/dag-json.git#egg=dag_json
git+https://github.com/snarfed/granary.git#egg=granary
git+https://github.com/snarfed/httpsig.git@HTTPSignatureAuth-sign-header#egg=httpsig
git+https://github.com/snarfed/lexrpc.git#egg=lexrpc
git+https://github.com/snarfed/oauth-dropins.git#egg=oauth_dropins
git+https://github.com/snarfed/requests_oauth2client.git@all#egg=requests_oauth2client

attrs==25.1.0
bases==0.3.0
beautifulsoup4==4.13.3
bech32==1.2.0
binapy==0.8.0
blinker==1.9.0
brevity==0.2.17
cachelib==0.13.0
cachetools==5.5.2
cbor2==5.6.5
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dag-cbor==0.3.3
dnspython==2.7.0
domain2idna==1.12.3
ecdsa==0.19.0
feedgen==1.0.0
feedparser==6.0.11
Flask==3.1.0
Flask-Caching==2.3.1
flask-gae-static==1.0
furl==2.1.4
gdata-python3==3.0.1
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-datastore==2.20.2
google-cloud-ndb==2.3.2
googleapis-common-protos==1.69.1
grapheme==0.6.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
html2text==2024.2.26
html5lib==1.1
humanfriendly==10.0
humanize==4.12.1
idna==3.10
iterators==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
jwskate==0.11.1
libipld==3.0.1
lxml==5.3.1
MarkupSafe==3.0.2
mf2py==2.0.1
mf2util==0.5.2
multiformats==0.3.1.post4
multiformats-config==0.3.1
oauthlib==3.2.2
orderedmultidict==1.0.1
pillow==11.1.0
pkce==1.0.3
praw==7.8.1
prawcore==2.4.0
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.9.0
pymediainfo==7.0.1
pymemcache==4.0.0
python-dateutil==2.9.0.post0
python-tumblpy==1.1.4
pytz==2025.1
pyyaml==6.0.2
redis==5.3.0b5
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
secp256k1==0.14.0
setuptools==76.0.0
sgmllib3k==1.0.0
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.6
tlslite-ng==0.8.2
tweepy==4.15.0
typing-validation==1.2.11.post4
typing_extensions==4.12.2
ujson==5.10.0
update-checker==0.18.0
urllib3==2.3.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
